Created ScroopsObject ancestor class and moved debug/error/warn functions there; implemented in other object
